renamed everything to match main version, added __init__ file
