+++ lib_new/__init__.py
@@ -1,0 +1,19 @@
+import os, copy, gc, math
+
+import matplotlib.pyplot as plt
+import numpy             as np
+import scipy             as sc
+import pandas            as pd
+import plotly.express    as px
+import plotly.io         as pio
+import gc
+from scipy.signal   import find_peaks
+from scipy.optimize import curve_fit
+from scipy.special  import erf
+from itertools      import product
+
+from .io_functions  import *
+from .wvf_functions import *
+from .cut_functions import *
+from .dec_functions import *
+from .ped_functions import *
